backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlmodel==0.0.16
sqlalchemy[asyncio]==2.0.29
aiosqlite==0.19.0
asyncpg==0.29.0
celery==5.3.6
redis==5.0.4
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
prometheus-client==0.20.0
transformers==4.39.3
accelerate==0.28.0
torch==2.2.1
torchvision==0.17.1
torchaudio==2.2.1
diffusers==0.27.2
safetensors==0.4.2
huggingface-hub==0.22.2
TTS==0.22.0
gTTS==2.5.1
pillow==10.3.0
ffmpeg-python==0.2.0
typer==0.15.1
langchain-openai==0.1.4
kokoro==0.9.4
google-api-python-client==2.122.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
youtube-transcript-api==0.6.1
